chore: Change variable type in constructor

+++ server/src/main/java/controller/AuctionMonitor.java
@@ -331,7 +331,7 @@ public class AuctionMonitor {
             ps.setString(2, winnerId);
             try (ResultSet rs = ps.executeQuery()) {
                 while (rs.next()) {
-                    walletDAO.unlockBalance(conn, rs.getString("bidder_id"), rs.getDouble("max_bid"));
+                    walletDAO.unlockBalance(conn, rs.getString("bidder_id"), rs.getLong("max_bid"));
                 }
             }
         }
